requirements.txt
<<<<<<< HEAD
groq>=0.11.0
streamlit>=1.33.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
youtube-transcript-api>=0.6.2
beautifulsoup4>=4.12.0
pypdf>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
PyYAML>=6.0.1
httpx>=0.27.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
torch>=2.1.0
transformers>=4.38.0
openai-whisper>=20231117
crewai>=0.76.0
litellm>=1.50.0
pdf2image>=1.17.0
pytesseract>=0.3.10
fpdf2>=2.7.0
# Optional (Windows may need extra build tools):
# face_recognition
=======
altair
pandas
streamlit
>>>>>>> d9be070ce79e63751534c32c5b34191f072134e2
